Provide convention that "join tables" have '_' prefix
These tables do not represent primary types but allow for
one-to-many or many-to-many relationships.

In the process, recognized that DataAttribute isn't
really its own type (but a way of representing attributes
for subclasses of `Data` in EAV format).

This closes #7 (with the exception of the inconsistency
that some data are represented in EAV and others in
separate sheets, but we are going to let that one go.